+++ src/youtube_utils/__init__.py
@@ -1,47 +1,7 @@
 from yt_dlp import YoutubeDL
 from .utils import Path
-import youtube_unofficial
 
-class WatchLater:
-    def __init__(self, data):
-        self.data = data
-
-    def raw_data(self):
-        return self.data
-
-    @staticmethod
-    def process_entry(entry):
-        # {
-        #     "description": null,
-
-        #     "channel_id": "UCNTVzV1InxHV-YR0fSajqPQ",
-
-        #     "uploader": "Supabase",
-        #     "uploader_id": "@Supabase",
-        #     "uploader_url": "https://www.youtube.com/@Supabase",
-
-        #     "release_timestamp": null,
-        #     "availability": null,
-
-        #     "live_status": null,
-        #     "channel_is_verified": null,
-        #     "__x_forwarded_for_ip": null
-        # },
-        thumb = max(entry["thumbnails"], key=lambda x: x["height"] * x["width"])
-        return {
-            "id": entry["id"],
-            "url": entry["url"],
-            "upload_date": entry["timestamp"],
-            "channel": entry["channel"],
-            "channel_url": entry["channel_url"],
-            "title": entry["title"],
-            "duration": entry["duration"],
-            "view_count": entry["view_count"],
-            "thumbnail": thumb["url"],
-        }
-
-    def items(self):
-        return list(map(self.process_entry, self.data["entries"]))
+from .watchlater import WatchLater
 
 
 class Youtube:
